% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.bands.R
\name{compute.bands}
\alias{compute.bands}
\title{Uniform confidence bands for QTE}
\usage{
compute.bands(G, Q, n.sim, n.sam, hh, alpha)
}
\description{
\code{compute.bands} is used to construct uniform confidence bands.
The function \code{make.band} calls this function to construct uniform bands.
}
\keyword{internal}
