% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depa.R
\name{depa}
\alias{depa}
\title{Epanechnikov kernel}
\usage{
depa(xx, loc = 0, scale = 1)
}
\arguments{
\item{xx}{the evaluation points.}

\item{loc}{the location parameter (normalized to be 0).}

\item{scale}{the scale parameter (normalized to be 1).}
}
\value{
values.
}
\description{
Epanechnikov kernel
}
\examples{
depa(0)
depa(seq(-1,1,by=0.05))

}
\keyword{internal}
