% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.band.qte.R
\name{print.band.qte}
\alias{print.band.qte}
\title{Print a band.qte object}
\usage{
\method{print}{band.qte}(x, ...)
}
\arguments{
\item{x}{object returned from \code{band.qte}.}

\item{...}{optional arguments.}
}
\description{
Print a band.qte object
}
