% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bw.qte.R
\name{print.bw.qte}
\alias{print.bw.qte}
\title{Print a bw.qte object}
\usage{
\method{print}{bw.qte}(x, ...)
}
\arguments{
\item{x}{object returned from \code{bandwidth.qte}.}

\item{...}{optional arguments.}
}
\description{
Print a bw.qte object
}
\keyword{internal}
