% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.qte.R
\name{print.qte}
\alias{print.qte}
\title{Print a qte object.}
\usage{
\method{print}{qte}(x, ...)
}
\arguments{
\item{x}{object returned from \code{rd.qte}.}

\item{...}{optional arguments.}
}
\description{
Print a qte object.
}
