% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.qte.R
\name{print.summary.qte}
\alias{print.summary.qte}
\title{Print a summary.qte object}
\usage{
\method{print}{summary.qte}(x, ...)
}
\arguments{
\item{x}{object returned from \code{summary.qte}.}

\item{...}{optional arguments.}
}
\description{
Print a summary.qte object
}
