% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.test.qte.R
\name{print.test.qte}
\alias{print.test.qte}
\title{Print a test.qte object}
\usage{
\method{print}{test.qte}(x, ...)
}
\arguments{
\item{x}{object returned from \code{test.qte}.}

\item{...}{optional arguments.}
}
\description{
Print a test.qte object
}
\keyword{internal}
